requirements.txt
beautifulsoup4==4.5.1
chardet==2.3.0
Django==1.10.3
django-authtools==1.5.0
django-crispy-forms==1.6.0
django-crontab==0.7.1
django-extensions==1.7.4
django-newsletter==0.6
django-tinymce==2.4.0
django-webtest==1.8.0
ldif3==3.1.1
pillow==3.4.2
python-card-me==0.9.3
python-dateutil==2.5.3
six==1.10.0
sorl-thumbnail==12.3
surlex==0.2.0
unicodecsv==0.14.1
waitress==1.0.1
WebOb==1.6.2
WebTest==2.0.23
